+++ includes/class.GFEwayAdmin.php
@@ -98,7 +98,7 @@ class GFEwayAdmin {
 		}
 
 		// need these PHP extensions too
-		$prereqs = array('libxml', 'SimpleXML', 'xmlwriter');
+		$prereqs = array('libxml', 'pcre', 'SimpleXML', 'xmlwriter');
 		$missing = array();
 		foreach ($prereqs as $ext) {
 			if (!extension_loaded($ext)) {
@@ -107,6 +107,12 @@ class GFEwayAdmin {
 		}
 		if (!empty($missing)) {
 			include GFEWAY_PLUGIN_ROOT . 'views/requires-extensions.php';
+		}
+
+		// and PCRE needs to be v8+ or we break! e.g. \K not present until v7.2 and some sites still use v6.6!
+		$pcre_min = '8';
+		if (defined('PCRE_VERSION') && version_compare(PCRE_VERSION, $pcre_min, '<')) {
+			include GFEWAY_PLUGIN_ROOT . 'views/requires-pcre.php';
 		}
 
 		// and of course, we need Gravity Forms