+++ changelog.md
@@ -5,6 +5,7 @@
 ### 2.2.2, soon...
 
 * fixed: no confirmation message after saving settings
+* changed: use the minified version of the eWAY Client Side Encryption script, unless `SCRIPT_DEBUG` is enabled
 
 ### 2.2.1, 2016-05-06